requirements.txt
streamlit==1.36.0
pandas==2.2.3
numpy==1.26.0
matplotlib==3.7.2
reportlab==4.2.0
google-generativeai>=0.8.3
